refactor(pom): extract shared PPTX layout enumeration helpers
extractThemeTokensFromPptx.ts の master/layout 列挙・非表示判定ロジックを
pptxLayoutEnumeration.ts へ切り出し、extractSlideMastersAsPptx (#907) と
列挙順を共有できるようにする。fixture ビルダーも testUtils/pptxFixture.ts
へ移動し、両テストで再利用する。
